requirements.txt
mock
pyserial==3.4
paho-mqtt
argparse
